chore: rename elementSelf to hostElement
