+++ packages/@lwc/engine-core/src/framework/base-lightning-element.ts
@@ -195,7 +195,7 @@ export interface LightningElement extends HTMLElementTheGoodParts, AccessibleEle
     constructor: LightningElementConstructor;
     template: ShadowRoot | null;
     refs: RefNodes | undefined;
-    elementSelf: Element;
+    hostElement: Element;
     render(): Template;
     connectedCallback?(): void;
     disconnectedCallback?(): void;
@@ -544,13 +544,13 @@ function warnIfInvokedDuringConstruction(vm: VM, methodOrPropName: string) {
         return vm.shadowRoot;
     },
 
-    get elementSelf(): Element {
+    get hostElement(): Element {
         const vm = getAssociatedVM(this);
 
         if (process.env.NODE_ENV !== 'production') {
             assert.isTrue(
                 vm.elm instanceof Element,
-                `this.elementSelf should be an Element, found: ${vm.elm}`
+                `this.hostElement should be an Element, found: ${vm.elm}`
             );
         }
