+++ packages/@lwc/integration-karma/test/component/LightningElement.hostElement/x/light/light.html
