+++ packages/@lwc/integration-karma/test/component/LightningElement.hostElement/x/shadow/shadow.js
@@ -3,6 +3,6 @@ import { LightningElement, api } from 'lwc';
 export default class Shadow extends LightningElement {
     @api
     getHostElement() {
-        return this.elementSelf;
+        return this.hostElement;
     }
 }
