+++ packages/@lwc/integration-karma/test/component/LightningElement.hostElement/x/wrapper/wrapper.js
