+++ packages/@lwc/integration-karma/test/component/properties/index.spec.js
@@ -14,7 +14,7 @@ const expectedEnumerableProps = [
     'dir',
     'dispatchEvent',
     'draggable',
-    'elementSelf',
+    'hostElement',
     'firstChild',
     'firstElementChild',
     'getAttribute',
